code cleanup

+++ src/sparrow/basename.c
@@ -1,3 +1,4 @@
+#include <u.h>
 #include <avian.h>
 #include <libgen.h>
 
@@ -9,7 +10,7 @@ usage(void) {
 
 int
 main(int argc, char *argv[]) {
-  size_t n;
+  uint n;
   char *p;
 
   ARGBEGIN(""){
@@ -21,7 +22,7 @@ main(int argc, char *argv[]) {
     usage();
   p = basename(argv[0]);
   if(argc == 2) {
-    n = strlen(p) - strlen(argv[1]);
+    n = strlen(p)-strlen(argv[1]);
     if(n > 0 && !strcmp(p+n, argv[1]))
       p[n] = '\0';
   }
